+++ CHANGELOG.md
@@ -1,7 +1,10 @@
 # Changelog
 
-## v0.0.2 - ?
+## v0.1.0 - ?
 
+- Add sops::create_value_in_vault to ensure a value exists in an encrypted sops file, and get a reference to it
+- Add sops::create_decrypted_value_reference to reference a value in a decrypted sops file
+- Add sops::create_decrypted_file_reference to decrypt a sops file with a reference
 
 ## v0.0.1 - 2025-12-27